test/e2e/pyproject.toml
[tool.poetry]
name = "e2e"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
supabase = {path = "/Users/filipecabaco/workspace/supabase-py"}
python-dotenv = "^1.0.1"
pytest = "^8.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
